requirements.txt
gradio_client==1.8.0
python-telegram-bot==21.4
python-dotenv==0.19.0
flask==3.0.2
openai==1.12.0
